backend/life_signatures.py
```python
from AI_models.life_signatures_model import LifeSignaturesModel
from backend.data_processing import processCosmologicalData

class LifeSignatures:
    def __init__(self):
        self.model = LifeSignaturesModel()
        self.data = processCosmologicalData()

    def identifyLifeSignatures(self):
        lifeSignatureData = []
        for planet in self.data:
            signature = self.model.predict(planet)
            lifeSignatureData.append({
                'planet': planet,
                'signature': signature
            })
        return lifeSignatureData
```